Adds github PAT for bumping poetry version.
To trigger github actions this MUST be ran as a user using a PAT, the
normal GITHUB_TOKEN does not have permission to trigger actions
(to prevent infinite loops).

We know we want to trigger a CI run after bumping the version, and
the version should NOT bump twice, so it's safe.